feat: add more purge tmdb actions
Added a new purge tmdb movie collections action and a new purge tmdb show alternate orderigns action. Also fixed some typos and grammar in the existing tmdb actions. I don't like how the actions are added to the list, but i also don't feel like writing it better, so it is what it is.